backend/requirements.txt
Flask
flask-cors
cryptography
python-dotenv
gunicorn